+++ commons-text/src/main/java/org/jdbi/commonstext/package-info.java
@@ -11,20 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.core.qualifier;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import org.jdbi.core.mapper.ColumnMapper;
-import org.jdbi.core.statement.StatementContext;
-
-// tag::columnmapper[]
-@Reversed // <1>
-public class ReversedStringMapper implements ColumnMapper<String> {
-    @Override
-    public String map(ResultSet r, int columnNumber, StatementContext ctx) throws SQLException {
-        return Reverser.reverse(r.getString(columnNumber));
-    }
-}
-// end::columnmapper[]
+/**
+ * <p>
+ * A template engine implementation using Apache Commons Text
+ * {@link org.apache.commons.text.StringSubstitutor}. This allows
+ * configurable variable delimiters in SQL templates, such as
+ * {@code ${foo}}, {@code <foo>}, or {@code %foo%}.
+ * </p>
+ */
+package org.jdbi.commonstext;