+++ core/src/main/java/org/jdbi/core/async/package-info.java
@@ -11,20 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.core.qualifier;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import org.jdbi.core.mapper.ColumnMapper;
-import org.jdbi.core.statement.StatementContext;
-
-// tag::columnmapper[]
-@Reversed // <1>
-public class ReversedStringMapper implements ColumnMapper<String> {
-    @Override
-    public String map(ResultSet r, int columnNumber, StatementContext ctx) throws SQLException {
-        return Reverser.reverse(r.getString(columnNumber));
-    }
-}
-// end::columnmapper[]
+/**
+ * <p>
+ * Provides asynchronous execution of Jdbi operations.
+ * The {@link org.jdbi.core.async.JdbiExecutor} wraps a
+ * {@link org.jdbi.core.Jdbi} instance and an {@link java.util.concurrent.Executor}
+ * to run callbacks asynchronously, returning
+ * {@link java.util.concurrent.CompletionStage} results.
+ * </p>
+ */
+package org.jdbi.core.async;