+++ core/src/main/java/org/jdbi/core/codec/package-info.java
@@ -11,20 +11,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.core.qualifier;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import org.jdbi.core.mapper.ColumnMapper;
-import org.jdbi.core.statement.StatementContext;
-
-// tag::columnmapper[]
-@Reversed // <1>
-public class ReversedStringMapper implements ColumnMapper<String> {
-    @Override
-    public String map(ResultSet r, int columnNumber, StatementContext ctx) throws SQLException {
-        return Reverser.reverse(r.getString(columnNumber));
-    }
-}
-// end::columnmapper[]
+/**
+ * <p>
+ * Bidirectional mapping between Java types and database columns.
+ * A {@link org.jdbi.core.codec.Codec} combines a
+ * {@link org.jdbi.core.mapper.ColumnMapper} and an
+ * {@link org.jdbi.core.argument.Argument} mapping function.
+ * The {@link org.jdbi.core.codec.CodecFactory} manages codec
+ * registration and lookup.
+ * </p>
+ */
+package org.jdbi.core.codec;