+++ core/src/main/java/org/jdbi/core/mapper/immutables/package-info.java
@@ -11,20 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.core.qualifier;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import org.jdbi.core.mapper.ColumnMapper;
-import org.jdbi.core.statement.StatementContext;
-
-// tag::columnmapper[]
-@Reversed // <1>
-public class ReversedStringMapper implements ColumnMapper<String> {
-    @Override
-    public String map(ResultSet r, int columnNumber, StatementContext ctx) throws SQLException {
-        return Reverser.reverse(r.getString(columnNumber));
-    }
-}
-// end::columnmapper[]
+/**
+ * <p>
+ * Support for mapping to and from
+ * <a href="https://immutables.github.io">Immutables</a> generated value
+ * types. Register value types through the
+ * {@link org.jdbi.core.mapper.immutables.JdbiImmutables} configuration
+ * class.
+ * </p>
+ */
+package org.jdbi.core.mapper.immutables;