+++ jackson2/src/main/java/org/jdbi/jackson2/package-info.java
@@ -11,20 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.core.qualifier;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import org.jdbi.core.mapper.ColumnMapper;
-import org.jdbi.core.statement.StatementContext;
-
-// tag::columnmapper[]
-@Reversed // <1>
-public class ReversedStringMapper implements ColumnMapper<String> {
-    @Override
-    public String map(ResultSet r, int columnNumber, StatementContext ctx) throws SQLException {
-        return Reverser.reverse(r.getString(columnNumber));
-    }
-}
-// end::columnmapper[]
+/**
+ * <p>
+ * JSON support using the
+ * <a href="https://github.com/FasterXML/jackson">Jackson 2</a> library.
+ * Install the {@link org.jdbi.jackson2.Jackson2Plugin} to provide
+ * Jackson-backed JSON serialization for {@code @Json} qualified types.
+ * </p>
+ */
+package org.jdbi.jackson2;