+++ postgis/src/main/java/org/jdbi/postgis/package-info.java
@@ -11,20 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.core.qualifier;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import org.jdbi.core.mapper.ColumnMapper;
-import org.jdbi.core.statement.StatementContext;
-
-// tag::columnmapper[]
-@Reversed // <1>
-public class ReversedStringMapper implements ColumnMapper<String> {
-    @Override
-    public String map(ResultSet r, int columnNumber, StatementContext ctx) throws SQLException {
-        return Reverser.reverse(r.getString(columnNumber));
-    }
-}
-// end::columnmapper[]
+/**
+ * <p>
+ * PostGIS support for Jdbi. Install the
+ * {@link org.jdbi.postgis.PostgisPlugin} to bind and map JTS
+ * geometry types such as {@code Point}, {@code Polygon}, and
+ * {@code LineString} to PostGIS columns.
+ * </p>
+ */
+package org.jdbi.postgis;